+++ ldaptor/test/test_inmemory.py
@@ -9,6 +9,11 @@ import six
 from ldaptor import inmemory, delta, testutil
 from ldaptor.protocols.ldap import distinguishedname, ldaperrors
 
+
+class SubclassEntry(inmemory.ReadOnlyInMemoryLDAPEntry):
+    pass
+
+
 class TestInMemoryDatabase(unittest.TestCase):
     def setUp(self):
         self.root = inmemory.ReadOnlyInMemoryLDAPEntry(
@@ -127,6 +132,15 @@ class TestInMemoryDatabase(unittest.TestCase):
             'objectClass': ['a'],
             'cn': 'foo',
             })
+
+    def test_addChild_subclass(self):
+        """
+        Adding child to ReadOnlyInMemoryLDAPEntry subclass instance
+        creates entry of the same class
+        """
+        entry = SubclassEntry(dn='dc=example,dc=com')
+        child = entry.addChild(rdn='ou=empty', attributes={'objectClass': ['a']})
+        self.assertIsInstance(child, SubclassEntry)
 
     def test_parent(self):
         self.assertEqual(self.foo.parent(), self.meta)
@@ -473,8 +487,8 @@ cn: foo
         def eb(fail):
             fail.trap(ldaperrors.LDAPNoSuchObject)
             self.failUnlessEqual(
-                str(fail.value),
-                'noSuchObject: ou=nonexisting,dc=example,dc=com')
+                fail.value.toWire(),
+                b'noSuchObject: ou=nonexisting,dc=example,dc=com')
         d.addCallbacks(testutil.mustRaise, eb)
         return d