+++ ldaptor/test/test_proxybase.py
@@ -105,10 +105,12 @@ class ProxyBase(unittest.TestCase):
         result code.is written to the transport.
         """
         server = self.createServer([pureldap.LDAPBindResponse(resultCode=0)])
-        server.dataReceived(str(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=4)))
+        server.dataReceived(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=4).toWire())
         server.reactor.advance(1)
-        self.assertEqual(server.transport.value(),
-                          str(pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=4)))
+        self.assertEqual(
+            server.transport.value(),
+            pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=4).toWire()
+        )
 
     def test_search(self):
         """
@@ -120,14 +122,28 @@ class ProxyBase(unittest.TestCase):
                                     pureldap.LDAPSearchResultEntry('cn=bar,dc=example,dc=com', [('b', ['c'])]),
                                     pureldap.LDAPSearchResultDone(ldaperrors.Success.resultCode)],
                                    )
-        server.dataReceived(str(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=2)))
-        server.dataReceived(str(pureldap.LDAPMessage(pureldap.LDAPSearchRequest(), id=3)))
+        server.dataReceived(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=2).toWire())
+        server.dataReceived(pureldap.LDAPMessage(pureldap.LDAPSearchRequest(), id=3).toWire())
         server.reactor.advance(1)
-        self.assertEqual(server.transport.value(),
-                          str(pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2))
-                          + str(pureldap.LDAPMessage(pureldap.LDAPSearchResultEntry('cn=foo,dc=example,dc=com', [('a', ['b'])]), id=3))
-                          + str(pureldap.LDAPMessage(pureldap.LDAPSearchResultEntry('cn=bar,dc=example,dc=com', [('b', ['c'])]), id=3))
-                          + str(pureldap.LDAPMessage(pureldap.LDAPSearchResultDone(ldaperrors.Success.resultCode), id=3)))
+        self.assertEqual(
+            server.transport.value(),
+            pureldap.LDAPMessage(
+                pureldap.LDAPBindResponse(resultCode=0),
+                id=2
+            ).toWire()
+            + pureldap.LDAPMessage(
+                pureldap.LDAPSearchResultEntry('cn=foo,dc=example,dc=com', [('a', ['b'])]),
+                id=3
+            ).toWire()
+            + pureldap.LDAPMessage(
+                pureldap.LDAPSearchResultEntry('cn=bar,dc=example,dc=com', [('b', ['c'])]),
+                id=3
+            ).toWire()
+            + pureldap.LDAPMessage(
+                pureldap.LDAPSearchResultDone(ldaperrors.Success.resultCode),
+                id=3
+            ).toWire()
+        )
 
     def test_unbind_clientUnbinds(self):
         """
@@ -135,19 +151,23 @@ class ProxyBase(unittest.TestCase):
         client signals its intent to unbind.
         """
         server = self.createServer([pureldap.LDAPBindResponse(resultCode=0)], [],)
-        server.dataReceived(str(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=2)))
+        server.dataReceived(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=2).toWire())
         server.reactor.advance(1)
         client = server.client
         client.assertSent(pureldap.LDAPBindRequest())
-        self.assertEqual(server.transport.value(),
-                          str(pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2)))
-        server.dataReceived(str(pureldap.LDAPMessage(pureldap.LDAPUnbindRequest(), id=3)))
+        self.assertEqual(
+            server.transport.value(),
+            pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2).toWire()
+        )
+        server.dataReceived(pureldap.LDAPMessage(pureldap.LDAPUnbindRequest(), id=3).toWire())
         server.connectionLost(error.ConnectionDone)
         server.reactor.advance(1)
         client.assertSent(pureldap.LDAPBindRequest(),
                           pureldap.LDAPUnbindRequest())
-        self.assertEqual(server.transport.value(),
-                          str(pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2)))
+        self.assertEqual(
+            server.transport.value(),
+            pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2).toWire()
+        )
 
     def test_unbind_clientEOF(self):
         """
@@ -155,20 +175,23 @@ class ProxyBase(unittest.TestCase):
         connection without sending an unbind request.
         """
         server = self.createServer([pureldap.LDAPBindResponse(resultCode=0)])
-        server.dataReceived(str(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=2)))
+        server.dataReceived(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=2).toWire())
         server.reactor.advance(1)
         client = server.client
         client.assertSent(pureldap.LDAPBindRequest())
         self.assertEqual(
             server.transport.value(),
-            str(pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2)))
+            pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2).toWire()
+        )
         server.connectionLost(error.ConnectionDone)
         server.reactor.advance(1)
         client.assertSent(
             pureldap.LDAPBindRequest(),
             'fake-unbind-by-LDAPClientTestDriver')
-        self.assertEqual(server.transport.value(),
-                          str(pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2)))
+        self.assertEqual(
+            server.transport.value(),
+            pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2).toWire()
+        )
 
     def test_intercepted_search_request(self):
         """
@@ -186,12 +209,20 @@ class ProxyBase(unittest.TestCase):
         server.responses = [
             pureldap.LDAPSearchResultEntry('cn=xyzzy,dc=example,dc=com', [('frobnitz', ['zork'])]),
             pureldap.LDAPSearchResultDone(ldaperrors.Success.resultCode)]
-        server.dataReceived(str(pureldap.LDAPMessage(pureldap.LDAPSearchRequest(), id=1)))
+        server.dataReceived(pureldap.LDAPMessage(pureldap.LDAPSearchRequest(), id=1).toWire())
         server.reactor.advance(1)
         self.assertEqual(len(server.clientTestDriver.sent), 0)
-        self.assertEqual(server.transport.value(),
-                          str(pureldap.LDAPMessage(pureldap.LDAPSearchResultEntry('cn=xyzzy,dc=example,dc=com', [('frobnitz', ['zork'])]), id=1))
-                          + str(pureldap.LDAPMessage(pureldap.LDAPSearchResultDone(ldaperrors.Success.resultCode), id=1)))
+        self.assertEqual(
+            server.transport.value(),
+            pureldap.LDAPMessage(
+                pureldap.LDAPSearchResultEntry('cn=xyzzy,dc=example,dc=com', [('frobnitz', ['zork'])]),
+                id=1
+            ).toWire()
+            + pureldap.LDAPMessage(
+                pureldap.LDAPSearchResultDone(ldaperrors.Success.resultCode),
+                id=1
+            ).toWire()
+        )
 
     def test_intercepted_search_response(self):
         """
@@ -203,15 +234,26 @@ class ProxyBase(unittest.TestCase):
                                     pureldap.LDAPSearchResultEntry('cn=bar,dc=example,dc=com', [('b', ['c'])]),
                                     pureldap.LDAPSearchResultDone(ldaperrors.Success.resultCode)],
                                    protocol=ResponseInterceptingProxy)
-        server.dataReceived(str(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=2)))
-        server.dataReceived(str(pureldap.LDAPMessage(pureldap.LDAPSearchRequest(), id=3)))
+        server.dataReceived(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=2).toWire())
+        server.dataReceived(pureldap.LDAPMessage(pureldap.LDAPSearchRequest(), id=3).toWire())
         server.reactor.advance(1)
         server.reactor.advance(5)
-        self.assertEqual(server.transport.value(),
-                          str(pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2))
-                          + str(pureldap.LDAPMessage(pureldap.LDAPSearchResultEntry('cn=foo,dc=example,dc=com', [('a', ['b']), ('frotz', ['xyzzy'])]), id=3))
-                          + str(pureldap.LDAPMessage(pureldap.LDAPSearchResultEntry('cn=bar,dc=example,dc=com', [('b', ['c']), ('frotz', ['xyzzy'])]), id=3))
-                          + str(pureldap.LDAPMessage(pureldap.LDAPSearchResultDone(ldaperrors.Success.resultCode), id=3)))
+        self.assertEqual(
+            server.transport.value(),
+            pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=0), id=2).toWire()
+            + pureldap.LDAPMessage(
+                pureldap.LDAPSearchResultEntry('cn=foo,dc=example,dc=com', [('a', ['b']), ('frotz', ['xyzzy'])]),
+                id=3
+            ).toWire()
+            + pureldap.LDAPMessage(
+                pureldap.LDAPSearchResultEntry('cn=bar,dc=example,dc=com', [('b', ['c']), ('frotz', ['xyzzy'])]),
+                id=3
+            ).toWire()
+            + pureldap.LDAPMessage(
+                pureldap.LDAPSearchResultDone(ldaperrors.Success.resultCode),
+                id=3
+            ).toWire()
+        )
 
     def test_cannot_connect_to_proxied_server_no_pending_requests(self):
         """
@@ -242,11 +284,13 @@ class ProxyBase(unittest.TestCase):
             clock=clock)
         self.assertEqual(connector, server.clientConnector)
         server.dataReceived(
-            str(pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=4)))
+            pureldap.LDAPMessage(pureldap.LDAPBindRequest(), id=4).toWire()
+        )
         server.reactor.advance(2)
         self.assertEqual(
             server.transport.value(),
-            str(pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=52), id=4)))
+            pureldap.LDAPMessage(pureldap.LDAPBindResponse(resultCode=52), id=4).toWire()
+        )
 
     def test_health_check_closes_connection_to_proxied_server(self):
         """
@@ -258,7 +302,7 @@ class ProxyBase(unittest.TestCase):
         message = pureldap.LDAPMessage(request, id=4)
         server = self.createServer()
         # Send a message, message is queued
-        server.dataReceived(str(message))
+        server.dataReceived(message.toWire())
         self.assertEqual(len(server.queuedRequests), 1)
         self.assertEqual(server.queuedRequests[0][0], request)
         # Lose connection, message is discarded